Update Haniel Netzach.py
Toby, when I tried to run this code, the Python Turtle graphic was running off the screen. I am not sure how to correct this issue, but I've made a small adjustment to avoid unnecessary import statements inside the loop and to use a loop counter i instead of the variable x. I have done this for all of the code so far. I am concerned that this code will not function properly.