Update Inputted Generated Mandala.py
Toby,

I am having a difficult time fixing the color-setting issue with this code. It might be best to have Prof. Keeley take a look at this. I tried editing the code on a private Python file, and that specific part of your code is giving a lot of errors. Please let me know if there is anything else I can do to help.