Update import turtle.py
I discovered an indentation error in your code which caused "bob" to be undefined. With a small edit to the indentation of your code, I was able to fix and run it successfully. It gives a nice Turtle graphic too! This is one of my favorite graphics out of all the code!